Extract sweep gradient tests to custom mod
